// src/app/pages/costs/components/costs-table/costs.interface.ts
export interface CostsInterface {
    nickname: string;
    usuarioauth: string;
    claveauth: string;
    idempresa: number;
    idrol: number;
    usuario: string;
    contrasena: string;
    nombre: string;
    email: string;
    telefono: string;
    costo: number;
    idstatususuario: number;
    emailsms: string;
    bfechainicial: boolean;
    fechainicial: string;
}